get_confluence_page_as_storage_format/requirements.txt
boto3==1.17.95
atlassian-python-api==3.11.0